eisenmann-backend/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-dotenv==1.4.2
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.0.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pyparsing==3.0.7
pytz==2021.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
